Mountain::ApplicationState::ApplicationState

Struct ApplicationState

pub struct ApplicationState {

    pub WorkSpaceFolders: Arc<Mutex<Vec<WorkSpaceFolderStateDTO>>>,
    pub WorkSpaceConfigurationPath: Arc<Mutex<Option<PathBuf>>>,
    pub IsTrusted: Arc<AtomicBool>,
    pub WindowState: Arc<Mutex<WindowStateDTO>>,
    pub ActiveDocumentURI: Arc<Mutex<Option<String>>>,
    pub Configuration: Arc<Mutex<MergedConfigurationStateDTO>>,
    pub GlobalMemento: Arc<Mutex<HashMap<String, Value>>>,
    pub GlobalMementoPath: PathBuf,
    pub WorkSpaceMemento: Arc<Mutex<HashMap<String, Value>>>,
    pub WorkSpaceMementoPath: Arc<Mutex<Option<PathBuf>>>,
    pub CommandRegistry: Arc<Mutex<HashMap<String, CommandHandler<Wry>>>>,
    pub LanguageProviders: Arc<Mutex<HashMap<u32, ProviderRegistrationDTO>>>,
    pub NextProviderHandle: Arc<AtomicU32>,
    pub ScannedExtensions: Arc<Mutex<HashMap<String, ExtensionDescriptionStateDTO>>>,
    pub EnabledProposedAPIs: Arc<Mutex<HashMap<String, Vec<String>>>>,
    pub ExtensionScanPaths: Arc<Mutex<Vec<PathBuf>>>,
    pub DiagnosticsMap: Arc<Mutex<HashMap<String, HashMap<String, Vec<MarkerDataDTO>>>>>,
    pub OpenDocuments: Arc<Mutex<HashMap<String, DocumentStateDTO>>>,
    pub OutputChannels: Arc<Mutex<HashMap<String, OutputChannelStateDTO>>>,
    pub ActiveTerminals: Arc<Mutex<HashMap<u64, Arc<Mutex<TerminalStateDTO>>>>>,
    pub NextTerminalIdentifier: Arc<AtomicU64>,
    pub ActiveWebViews: Arc<Mutex<HashMap<String, WebViewStateDTO>>>,
    pub ActiveCustomDocuments: Arc<Mutex<HashMap<String, CustomDocumentStateDTO>>>,
    pub ActiveStatusBarItems: Arc<Mutex<HashMap<String, StatusBarEntryDTO>>>,
    pub ActiveTreeViews: Arc<Mutex<HashMap<String, TreeViewStateDTO>>>,
    pub SourceControlManagementProviders: Arc<Mutex<HashMap<u32, SourceControlManagementProviderDTO>>>,
    pub SourceControlManagementGroups: Arc<Mutex<HashMap<u32, HashMap<String, SourceControlManagementGroupDTO>>>>,
    pub SourceControlManagementResources: Arc<Mutex<HashMap<u32, HashMap<String, Vec<SourceControlManagementResourceDTO>>>>>,
    pub NextSourceControlManagementProviderHandle: Arc<AtomicU32>,
    pub PendingUserInterfaceRequests: Arc<Mutex<HashMap<String, Sender<Result<Value, CommonError>>>>>,

}

The central, shared, thread-safe state for the entire Mountain application.

Fields
WorkSpaceFolders: Arc<Mutex<Vec<WorkSpaceFolderStateDTO>>>WorkSpaceConfigurationPath: Arc<Mutex<Option<PathBuf>>>IsTrusted: Arc<AtomicBool>WindowState: Arc<Mutex<WindowStateDTO>>ActiveDocumentURI: Arc<Mutex<Option<String>>>Configuration: Arc<Mutex<MergedConfigurationStateDTO>>GlobalMemento: Arc<Mutex<HashMap<String, Value>>>GlobalMementoPath: PathBufWorkSpaceMemento: Arc<Mutex<HashMap<String, Value>>>WorkSpaceMementoPath: Arc<Mutex<Option<PathBuf>>>CommandRegistry: Arc<Mutex<HashMap<String, CommandHandler<Wry>>>>LanguageProviders: Arc<Mutex<HashMap<u32, ProviderRegistrationDTO>>>NextProviderHandle: Arc<AtomicU32>ScannedExtensions: Arc<Mutex<HashMap<String, ExtensionDescriptionStateDTO>>>EnabledProposedAPIs: Arc<Mutex<HashMap<String, Vec<String>>>>ExtensionScanPaths: Arc<Mutex<Vec<PathBuf>>>DiagnosticsMap: Arc<Mutex<HashMap<String, HashMap<String, Vec<MarkerDataDTO>>>>>OpenDocuments: Arc<Mutex<HashMap<String, DocumentStateDTO>>>OutputChannels: Arc<Mutex<HashMap<String, OutputChannelStateDTO>>>ActiveTerminals: Arc<Mutex<HashMap<u64, Arc<Mutex<TerminalStateDTO>>>>>NextTerminalIdentifier: Arc<AtomicU64>ActiveWebViews: Arc<Mutex<HashMap<String, WebViewStateDTO>>>ActiveCustomDocuments: Arc<Mutex<HashMap<String, CustomDocumentStateDTO>>>ActiveStatusBarItems: Arc<Mutex<HashMap<String, StatusBarEntryDTO>>>ActiveTreeViews: Arc<Mutex<HashMap<String, TreeViewStateDTO>>>SourceControlManagementProviders: Arc<Mutex<HashMap<u32, SourceControlManagementProviderDTO>>>SourceControlManagementGroups: Arc<Mutex<HashMap<u32, HashMap<String, SourceControlManagementGroupDTO>>>>SourceControlManagementResources: Arc<Mutex<HashMap<u32, HashMap<String, Vec<SourceControlManagementResourceDTO>>>>>NextSourceControlManagementProviderHandle: Arc<AtomicU32>PendingUserInterfaceRequests: Arc<Mutex<HashMap<String, Sender<Result<Value, CommonError>>>>>
Implementations

impl ApplicationState

pub fn GetWorkSpaceIdentifier(&self) -> Result<String, CommonError>

Generates a unique, filesystem-safe identifier for the current workspace. Returns “NO_WORKSPACE” if no folder or configuration is open.

pub fn GetNextProviderHandle(&self) -> u32

Returns the next available unique identifier for a language provider registration.

pub fn GetNextTerminalIdentifier(&self) -> u64

Returns the next available unique identifier for a terminal instance.

pub fn GetNextSourceControlManagementProviderHandle(&self) -> u32

Returns the next available unique identifier for an SCM provider instance.

pub fn UpdateWorkSpaceMementoPathAndReload( &self, ApplicationDataDirectory: &Path, ) -> Result<(), CommonError>

Updates the path to the workspace memento file and reloads its content from disk. This should be called when a workspace is opened or changed.