Mountain::ApplicationState

Module DTO

Defines all DTOs used to represent the various components of the application state.

ApplicationState DTO Module

This module aggregates and re-exports all Data Transfer Objects (DTOs) that are used to represent the various components of the application’s central state.

Modules

CustomDocumentStateDTO

CustomDocumentStateDTO

DocumentStateDTO

DocumentStateDTO

ExtensionDescriptionStateDTO

ExtensionDescriptionStateDTO

MarkerDataDTO

MarkerDataDTO

MergedConfigurationStateDTO

MergedConfigurationStateDTO

OutputChannelStateDTO

OutputChannelStateDTO

ProviderRegistrationDTO

ProviderRegistrationDTO

RPCModelContentChangeDTO

RPCModelContentChangeDTO

TerminalStateDTO

TerminalStateDTO

TreeViewStateDTO

TreeViewStateDTO

WebViewStateDTO

WebViewStateDTO

WindowStateDTO

WindowStateDTO

WorkSpaceFolderStateDTO

WorkSpaceFolderStateDTO