Mountain::Environment::UserInterface::DTO

Module FileFilterDTO

FileFilterDTO

Defines the Data Transfer Object for a single file filter in a native file dialog.

Structs

FileFilterDTO

A serializable struct that represents a single selectable filter in a file dialog’s “files of type” dropdown.