Mountain::Environment::UserInterface

Module UserInterfaceProvider

UserInterfaceProvider Trait

Defines the abstract service trait for all modal user interface interactions.

Traits

UserInterfaceProvider

An abstract service contract for an environment component that can perform UI interactions that require user input, such as showing dialogs, messages, and quick pick menus.