Mountain::Vine::Generated::cocoon_service_client

Struct CocoonServiceClient

pub struct CocoonServiceClient<T> {
    inner: Grpc<T>,
}

Service running on the Cocoon sidecar, listening for requests from Mountain.

Fields
inner: Grpc<T>
Implementations

impl CocoonServiceClient<Channel>

pub async fn connect<D>(dst: D) -> Result<Self, Error>
where D: TryInto<Endpoint>, D::Error: Into<StdError>,

Attempt to create a new client by connecting to a given endpoint.

impl<T> CocoonServiceClient<T>
where T: GrpcService<Body>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F, ) -> CocoonServiceClient<InterceptedService<T, F>>
where F: Interceptor, T::ResponseBody: Default, T: Service<Request<Body>, Response = Response<<T as GrpcService<Body>>::ResponseBody>>, <T as Service<Request<Body>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn process_mountain_request( &mut self, request: impl IntoRequest<GenericRequest>, ) -> Result<Response<GenericResponse>, Status>

A generic request-response method for Mountain to call a function on Cocoon.

pub async fn send_mountain_notification( &mut self, request: impl IntoRequest<GenericNotification>, ) -> Result<Response<Empty>, Status>

A generic fire-and-forget method for Mountain to send a notification to Cocoon.

pub async fn cancel_operation( &mut self, request: impl IntoRequest<CancelOperationRequest>, ) -> Result<Response<Empty>, Status>

A method for Mountain to request that Cocoon cancel a long-running operation.

pub async fn initial_handshake( &mut self, request: impl IntoRequest<Empty>, ) -> Result<Response<Empty>, Status>

Handshake - Called by Cocoon to signal readiness

pub async fn init_extension_host( &mut self, request: impl IntoRequest<InitExtensionHostRequest>, ) -> Result<Response<Empty>, Status>

Initialize Extension Host - Mountain sends initialization data to Cocoon

pub async fn register_command( &mut self, request: impl IntoRequest<RegisterCommandRequest>, ) -> Result<Response<Empty>, Status>

Register Command - Cocoon registers an extension command

pub async fn execute_contributed_command( &mut self, request: impl IntoRequest<ExecuteCommandRequest>, ) -> Result<Response<ExecuteCommandResponse>, Status>

Execute Contributed Command - Mountain executes an extension command

pub async fn unregister_command( &mut self, request: impl IntoRequest<UnregisterCommandRequest>, ) -> Result<Response<Empty>, Status>

Unregister Command - Unregister a previously registered command

pub async fn register_hover_provider( &mut self, request: impl IntoRequest<RegisterProviderRequest>, ) -> Result<Response<Empty>, Status>

Register Hover Provider - Register a hover provider

pub async fn provide_hover( &mut self, request: impl IntoRequest<ProvideHoverRequest>, ) -> Result<Response<ProvideHoverResponse>, Status>

Provide Hover - Request hover information

pub async fn register_completion_item_provider( &mut self, request: impl IntoRequest<RegisterProviderRequest>, ) -> Result<Response<Empty>, Status>

Register Completion Item Provider - Register a completion provider

pub async fn provide_completion_items( &mut self, request: impl IntoRequest<ProvideCompletionItemsRequest>, ) -> Result<Response<ProvideCompletionItemsResponse>, Status>

Provide Completion Items - Request completion items

pub async fn register_definition_provider( &mut self, request: impl IntoRequest<RegisterProviderRequest>, ) -> Result<Response<Empty>, Status>

Register Definition Provider - Register a definition provider

pub async fn provide_definition( &mut self, request: impl IntoRequest<ProvideDefinitionRequest>, ) -> Result<Response<ProvideDefinitionResponse>, Status>

Provide Definition - Request definition location

pub async fn register_reference_provider( &mut self, request: impl IntoRequest<RegisterProviderRequest>, ) -> Result<Response<Empty>, Status>

Register Reference Provider - Register a reference provider

pub async fn provide_references( &mut self, request: impl IntoRequest<ProvideReferencesRequest>, ) -> Result<Response<ProvideReferencesResponse>, Status>

Provide References - Request references

pub async fn register_code_actions_provider( &mut self, request: impl IntoRequest<RegisterProviderRequest>, ) -> Result<Response<Empty>, Status>

Register Code Actions Provider - Register code actions provider

pub async fn provide_code_actions( &mut self, request: impl IntoRequest<ProvideCodeActionsRequest>, ) -> Result<Response<ProvideCodeActionsResponse>, Status>

Provide Code Actions - Request code actions

pub async fn register_document_highlight_provider( &mut self, request: impl IntoRequest<RegisterProviderRequest>, ) -> Result<Response<Empty>, Status>

Register Document Highlight Provider

pub async fn provide_document_highlights( &mut self, request: impl IntoRequest<ProvideDocumentHighlightsRequest>, ) -> Result<Response<ProvideDocumentHighlightsResponse>, Status>

Provide Document Highlights

pub async fn register_document_symbol_provider( &mut self, request: impl IntoRequest<RegisterProviderRequest>, ) -> Result<Response<Empty>, Status>

Register Document Symbol Provider

pub async fn provide_document_symbols( &mut self, request: impl IntoRequest<ProvideDocumentSymbolsRequest>, ) -> Result<Response<ProvideDocumentSymbolsResponse>, Status>

Provide Document Symbols

pub async fn register_workspace_symbol_provider( &mut self, request: impl IntoRequest<RegisterProviderRequest>, ) -> Result<Response<Empty>, Status>

Register Workspace Symbol Provider

pub async fn provide_workspace_symbols( &mut self, request: impl IntoRequest<ProvideWorkspaceSymbolsRequest>, ) -> Result<Response<ProvideWorkspaceSymbolsResponse>, Status>

Provide Workspace Symbols

pub async fn register_rename_provider( &mut self, request: impl IntoRequest<RegisterProviderRequest>, ) -> Result<Response<Empty>, Status>

Register Rename Provider

pub async fn provide_rename_edits( &mut self, request: impl IntoRequest<ProvideRenameEditsRequest>, ) -> Result<Response<ProvideRenameEditsResponse>, Status>

Provide Rename Edits

pub async fn register_document_formatting_provider( &mut self, request: impl IntoRequest<RegisterProviderRequest>, ) -> Result<Response<Empty>, Status>

Register Document Formatting Provider

pub async fn provide_document_formatting( &mut self, request: impl IntoRequest<ProvideDocumentFormattingRequest>, ) -> Result<Response<ProvideDocumentFormattingResponse>, Status>

Provide Document Formatting

pub async fn register_document_range_formatting_provider( &mut self, request: impl IntoRequest<RegisterProviderRequest>, ) -> Result<Response<Empty>, Status>

Register Document Range Formatting Provider

pub async fn provide_document_range_formatting( &mut self, request: impl IntoRequest<ProvideDocumentRangeFormattingRequest>, ) -> Result<Response<ProvideDocumentRangeFormattingResponse>, Status>

Provide Document Range Formatting

pub async fn register_on_type_formatting_provider( &mut self, request: impl IntoRequest<RegisterOnTypeFormattingProviderRequest>, ) -> Result<Response<Empty>, Status>

Register On Type Formatting Provider

pub async fn provide_on_type_formatting( &mut self, request: impl IntoRequest<ProvideOnTypeFormattingRequest>, ) -> Result<Response<ProvideOnTypeFormattingResponse>, Status>

Provide On Type Formatting

pub async fn register_signature_help_provider( &mut self, request: impl IntoRequest<RegisterSignatureHelpProviderRequest>, ) -> Result<Response<Empty>, Status>

Register Signature Help Provider

pub async fn provide_signature_help( &mut self, request: impl IntoRequest<ProvideSignatureHelpRequest>, ) -> Result<Response<ProvideSignatureHelpResponse>, Status>

Provide Signature Help

pub async fn register_code_lens_provider( &mut self, request: impl IntoRequest<RegisterProviderRequest>, ) -> Result<Response<Empty>, Status>

Register Code Lens Provider

pub async fn provide_code_lenses( &mut self, request: impl IntoRequest<ProvideCodeLensesRequest>, ) -> Result<Response<ProvideCodeLensesResponse>, Status>

Provide Code Lenses

pub async fn register_folding_range_provider( &mut self, request: impl IntoRequest<RegisterProviderRequest>, ) -> Result<Response<Empty>, Status>

Register Folding Range Provider

pub async fn provide_folding_ranges( &mut self, request: impl IntoRequest<ProvideFoldingRangesRequest>, ) -> Result<Response<ProvideFoldingRangesResponse>, Status>

Provide Folding Ranges

pub async fn register_selection_range_provider( &mut self, request: impl IntoRequest<RegisterProviderRequest>, ) -> Result<Response<Empty>, Status>

Register Selection Range Provider

pub async fn provide_selection_ranges( &mut self, request: impl IntoRequest<ProvideSelectionRangesRequest>, ) -> Result<Response<ProvideSelectionRangesResponse>, Status>

Provide Selection Ranges

pub async fn register_semantic_tokens_provider( &mut self, request: impl IntoRequest<RegisterSemanticTokensProviderRequest>, ) -> Result<Response<Empty>, Status>

Register Semantic Tokens Provider

pub async fn provide_semantic_tokens_full( &mut self, request: impl IntoRequest<ProvideSemanticTokensRequest>, ) -> Result<Response<ProvideSemanticTokensResponse>, Status>

Provide Semantic Tokens Full

pub async fn register_inlay_hints_provider( &mut self, request: impl IntoRequest<RegisterProviderRequest>, ) -> Result<Response<Empty>, Status>

Register Inlay Hints Provider

pub async fn provide_inlay_hints( &mut self, request: impl IntoRequest<ProvideInlayHintsRequest>, ) -> Result<Response<ProvideInlayHintsResponse>, Status>

Provide Inlay Hints

pub async fn register_type_hierarchy_provider( &mut self, request: impl IntoRequest<RegisterProviderRequest>, ) -> Result<Response<Empty>, Status>

Register Type Hierarchy Provider

pub async fn provide_type_hierarchy_supertypes( &mut self, request: impl IntoRequest<ProvideTypeHierarchyRequest>, ) -> Result<Response<ProvideTypeHierarchyResponse>, Status>

Provide Type Hierarchy Supertypes

pub async fn provide_type_hierarchy_subtypes( &mut self, request: impl IntoRequest<ProvideTypeHierarchyRequest>, ) -> Result<Response<ProvideTypeHierarchyResponse>, Status>

Provide Type Hierarchy Subtypes

pub async fn register_call_hierarchy_provider( &mut self, request: impl IntoRequest<RegisterProviderRequest>, ) -> Result<Response<Empty>, Status>

Register Call Hierarchy Provider

pub async fn provide_call_hierarchy_incoming_calls( &mut self, request: impl IntoRequest<ProvideCallHierarchyRequest>, ) -> Result<Response<ProvideCallHierarchyResponse>, Status>

Provide Call Hierarchy Incoming Calls

pub async fn provide_call_hierarchy_outgoing_calls( &mut self, request: impl IntoRequest<ProvideCallHierarchyRequest>, ) -> Result<Response<ProvideCallHierarchyResponse>, Status>

Provide Call Hierarchy Outgoing Calls

pub async fn register_linked_editing_range_provider( &mut self, request: impl IntoRequest<RegisterProviderRequest>, ) -> Result<Response<Empty>, Status>

Register Linked Editing Range Provider

pub async fn provide_linked_editing_ranges( &mut self, request: impl IntoRequest<ProvideLinkedEditingRangesRequest>, ) -> Result<Response<ProvideLinkedEditingRangesResponse>, Status>

Provide Linked Editing Ranges

pub async fn show_quick_pick( &mut self, request: impl IntoRequest<ShowQuickPickRequest>, ) -> Result<Response<ShowQuickPickResponse>, Status>

Show Quick Pick

pub async fn show_input_box( &mut self, request: impl IntoRequest<ShowInputBoxRequest>, ) -> Result<Response<ShowInputBoxResponse>, Status>

Show Input Box

pub async fn show_progress( &mut self, request: impl IntoRequest<ShowProgressRequest>, ) -> Result<Response<ShowProgressResponse>, Status>

Show Progress

pub async fn report_progress( &mut self, request: impl IntoRequest<ReportProgressRequest>, ) -> Result<Response<Empty>, Status>

Report Progress

pub async fn post_webview_message( &mut self, request: impl IntoRequest<PostWebviewMessageRequest>, ) -> Result<Response<Empty>, Status>

Post Message to Webview

pub async fn dispose_webview_panel( &mut self, request: impl IntoRequest<DisposeWebviewPanelRequest>, ) -> Result<Response<Empty>, Status>

Dispose Webview Panel

pub async fn open_external( &mut self, request: impl IntoRequest<OpenExternalRequest>, ) -> Result<Response<Empty>, Status>

Open External URI

pub async fn delete_file( &mut self, request: impl IntoRequest<DeleteFileRequest>, ) -> Result<Response<Empty>, Status>

Delete File

pub async fn rename_file( &mut self, request: impl IntoRequest<RenameFileRequest>, ) -> Result<Response<Empty>, Status>

Rename File

pub async fn copy_file( &mut self, request: impl IntoRequest<CopyFileRequest>, ) -> Result<Response<Empty>, Status>

Copy File

pub async fn create_directory( &mut self, request: impl IntoRequest<CreateDirectoryRequest>, ) -> Result<Response<Empty>, Status>

Create Directory

pub async fn create_output_channel( &mut self, request: impl IntoRequest<CreateOutputChannelRequest>, ) -> Result<Response<CreateOutputChannelResponse>, Status>

Create Output Channel

pub async fn append_output( &mut self, request: impl IntoRequest<AppendOutputRequest>, ) -> Result<Response<Empty>, Status>

Append to Output Channel

pub async fn clear_output( &mut self, request: impl IntoRequest<ClearOutputRequest>, ) -> Result<Response<Empty>, Status>

Clear Output Channel

pub async fn show_output( &mut self, request: impl IntoRequest<ShowOutputRequest>, ) -> Result<Response<Empty>, Status>

Show Output Channel

pub async fn dispose_output( &mut self, request: impl IntoRequest<DisposeOutputRequest>, ) -> Result<Response<Empty>, Status>

Dispose Output Channel

pub async fn register_task_provider( &mut self, request: impl IntoRequest<RegisterTaskProviderRequest>, ) -> Result<Response<Empty>, Status>

Register Task Provider

pub async fn execute_task( &mut self, request: impl IntoRequest<ExecuteTaskRequest>, ) -> Result<Response<ExecuteTaskResponse>, Status>

Execute Task

pub async fn terminate_task( &mut self, request: impl IntoRequest<TerminateTaskRequest>, ) -> Result<Response<Empty>, Status>

Terminate Task

pub async fn get_authentication_session( &mut self, request: impl IntoRequest<GetAuthenticationSessionRequest>, ) -> Result<Response<GetAuthenticationSessionResponse>, Status>

Get Authentication Session

pub async fn register_authentication_provider( &mut self, request: impl IntoRequest<RegisterAuthenticationProviderRequest>, ) -> Result<Response<Empty>, Status>

Register Authentication Provider

pub async fn stop_debugging( &mut self, request: impl IntoRequest<StopDebuggingRequest>, ) -> Result<Response<Empty>, Status>

Stop Debugging

pub async fn get_extension( &mut self, request: impl IntoRequest<GetExtensionRequest>, ) -> Result<Response<GetExtensionResponse>, Status>

Get Extension

pub async fn get_all_extensions( &mut self, request: impl IntoRequest<Empty>, ) -> Result<Response<GetAllExtensionsResponse>, Status>

Get All Extensions

pub async fn resize_terminal( &mut self, request: impl IntoRequest<ResizeTerminalRequest>, ) -> Result<Response<Empty>, Status>

Resize Terminal

pub async fn get_configuration( &mut self, request: impl IntoRequest<GetConfigurationRequest>, ) -> Result<Response<GetConfigurationResponse>, Status>

Get Configuration Value

pub async fn show_text_document( &mut self, request: impl IntoRequest<ShowTextDocumentRequest>, ) -> Result<Response<ShowTextDocumentResponse>, Status>

Show Text Document - Open a text document

pub async fn show_information_message( &mut self, request: impl IntoRequest<ShowMessageRequest>, ) -> Result<Response<ShowMessageResponse>, Status>

Show Information Message - Display an info message

pub async fn show_warning_message( &mut self, request: impl IntoRequest<ShowMessageRequest>, ) -> Result<Response<ShowMessageResponse>, Status>

Show Warning Message - Display a warning message

pub async fn show_error_message( &mut self, request: impl IntoRequest<ShowMessageRequest>, ) -> Result<Response<ShowMessageResponse>, Status>

Show Error Message - Display an error message

pub async fn create_status_bar_item( &mut self, request: impl IntoRequest<CreateStatusBarItemRequest>, ) -> Result<Response<CreateStatusBarItemResponse>, Status>

Create Status Bar Item - Create a status bar item

pub async fn set_status_bar_text( &mut self, request: impl IntoRequest<SetStatusBarTextRequest>, ) -> Result<Response<Empty>, Status>

Set Status Bar Text - Set status bar text

pub async fn create_webview_panel( &mut self, request: impl IntoRequest<CreateWebviewPanelRequest>, ) -> Result<Response<CreateWebviewPanelResponse>, Status>

Create Webview Panel - Create a new webview panel

pub async fn set_webview_html( &mut self, request: impl IntoRequest<SetWebviewHtmlRequest>, ) -> Result<Response<Empty>, Status>

Set Webview HTML - Update webview HTML content

pub async fn on_did_receive_message( &mut self, request: impl IntoRequest<OnDidReceiveMessageRequest>, ) -> Result<Response<Empty>, Status>

On Did Receive Message - Receive message from webview

pub async fn read_file( &mut self, request: impl IntoRequest<ReadFileRequest>, ) -> Result<Response<ReadFileResponse>, Status>

Read File - Read file contents

pub async fn write_file( &mut self, request: impl IntoRequest<WriteFileRequest>, ) -> Result<Response<Empty>, Status>

Write File - Write file contents

pub async fn stat( &mut self, request: impl IntoRequest<StatRequest>, ) -> Result<Response<StatResponse>, Status>

Stat - Get file metadata

pub async fn readdir( &mut self, request: impl IntoRequest<ReaddirRequest>, ) -> Result<Response<ReaddirResponse>, Status>

Read Directory - List directory contents

pub async fn watch_file( &mut self, request: impl IntoRequest<WatchFileRequest>, ) -> Result<Response<Empty>, Status>

Watch File - Watch file for changes

pub async fn find_files( &mut self, request: impl IntoRequest<FindFilesRequest>, ) -> Result<Response<FindFilesResponse>, Status>

Find Files - Search for files

pub async fn find_text_in_files( &mut self, request: impl IntoRequest<FindTextInFilesRequest>, ) -> Result<Response<FindTextInFilesResponse>, Status>

Find Text in Files - Search for text across files

pub async fn open_document( &mut self, request: impl IntoRequest<OpenDocumentRequest>, ) -> Result<Response<OpenDocumentResponse>, Status>

Open Document - Open a document

pub async fn save_all( &mut self, request: impl IntoRequest<SaveAllRequest>, ) -> Result<Response<SaveAllResponse>, Status>

Save All - Save all open documents

pub async fn apply_edit( &mut self, request: impl IntoRequest<ApplyEditRequest>, ) -> Result<Response<ApplyEditResponse>, Status>

Apply Edit - Apply a text edit to a document

pub async fn update_configuration( &mut self, request: impl IntoRequest<UpdateConfigurationRequest>, ) -> Result<Response<Empty>, Status>

Update Configuration - Notify of configuration changes

pub async fn update_workspace_folders( &mut self, request: impl IntoRequest<UpdateWorkspaceFoldersRequest>, ) -> Result<Response<Empty>, Status>

Update Workspace Folders - Update workspace folders

pub async fn open_terminal( &mut self, request: impl IntoRequest<OpenTerminalRequest>, ) -> Result<Response<Empty>, Status>

Open Terminal - Open a new terminal

pub async fn terminal_input( &mut self, request: impl IntoRequest<TerminalInputRequest>, ) -> Result<Response<Empty>, Status>

Terminal Input - Send input to terminal

pub async fn close_terminal( &mut self, request: impl IntoRequest<CloseTerminalRequest>, ) -> Result<Response<Empty>, Status>

Close Terminal - Close a terminal

pub async fn accept_terminal_opened( &mut self, request: impl IntoRequest<TerminalOpenedNotification>, ) -> Result<Response<Empty>, Status>

Accept Terminal Opened - Notification: Terminal opened

pub async fn accept_terminal_closed( &mut self, request: impl IntoRequest<TerminalClosedNotification>, ) -> Result<Response<Empty>, Status>

Accept Terminal Closed - Notification: Terminal closed

pub async fn accept_terminal_process_id( &mut self, request: impl IntoRequest<TerminalProcessIdNotification>, ) -> Result<Response<Empty>, Status>

Accept Terminal Process ID - Notification: Terminal process ID

pub async fn accept_terminal_process_data( &mut self, request: impl IntoRequest<TerminalDataNotification>, ) -> Result<Response<Empty>, Status>

Accept Terminal Process Data - Notification: Terminal output

pub async fn register_tree_view_provider( &mut self, request: impl IntoRequest<RegisterTreeViewProviderRequest>, ) -> Result<Response<Empty>, Status>

Register Tree View Provider - Register a tree view provider

pub async fn get_tree_children( &mut self, request: impl IntoRequest<GetTreeChildrenRequest>, ) -> Result<Response<GetTreeChildrenResponse>, Status>

Get Tree Children - Request tree view children

pub async fn register_scm_provider( &mut self, request: impl IntoRequest<RegisterScmProviderRequest>, ) -> Result<Response<Empty>, Status>

Register SCM Provider - Register source control provider

pub async fn update_scm_group( &mut self, request: impl IntoRequest<UpdateScmGroupRequest>, ) -> Result<Response<Empty>, Status>

Update SCM Group - Update SCM group

pub async fn git_exec( &mut self, request: impl IntoRequest<GitExecRequest>, ) -> Result<Response<GitExecResponse>, Status>

Execute Git - Execute git command

pub async fn register_debug_adapter( &mut self, request: impl IntoRequest<RegisterDebugAdapterRequest>, ) -> Result<Response<Empty>, Status>

Register Debug Adapter - Register debug adapter

pub async fn start_debugging( &mut self, request: impl IntoRequest<StartDebuggingRequest>, ) -> Result<Response<StartDebuggingResponse>, Status>

Start Debugging - Start debug session

pub async fn participate_in_save( &mut self, request: impl IntoRequest<ParticipateInSaveRequest>, ) -> Result<Response<ParticipateInSaveResponse>, Status>

Participate in Save - Extension participates in save

pub async fn get_secret( &mut self, request: impl IntoRequest<GetSecretRequest>, ) -> Result<Response<GetSecretResponse>, Status>

Get Secret - Retrieve a secret from storage

pub async fn store_secret( &mut self, request: impl IntoRequest<StoreSecretRequest>, ) -> Result<Response<Empty>, Status>

Store Secret - Store a secret in storage

pub async fn delete_secret( &mut self, request: impl IntoRequest<DeleteSecretRequest>, ) -> Result<Response<Empty>, Status>

Delete Secret - Delete a secret from storage