Mountain::Vine::Generated::mountain_service_server

Struct MountainServiceServer

pub struct MountainServiceServer<T> {
    inner: Arc<T>,
    accept_compression_encodings: EnabledCompressionEncodings,
    send_compression_encodings: EnabledCompressionEncodings,
    max_decoding_message_size: Option<usize>,
    max_encoding_message_size: Option<usize>,
}

Service running on the Mountain host, listening for requests from Cocoon.

Fields
inner: Arc<T>accept_compression_encodings: EnabledCompressionEncodingssend_compression_encodings: EnabledCompressionEncodingsmax_decoding_message_size: Option<usize>max_encoding_message_size: Option<usize>
Implementations

impl<T> MountainServiceServer<T>

pub fn new(inner: T) -> Self

pub fn from_arc(inner: Arc<T>) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F, ) -> InterceptedService<Self, F>
where F: Interceptor,

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing requests with the given encoding.

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress responses with the given encoding, if the client supports it.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX