Crate Mountain

Mountain Crate Library

This file conceptually represents the library root for the Mountain application, declaring all of its major internal components. This allows the Binary.rs file to have a clean entry point that orchestrates these components.

Modules

ApplicationState

ApplicationState Module

Binary

Mountain Binary Entry Point

Command

Command Module

Environment

Environment Module

ExtensionManagement

ExtensionManagement Module

FileSystem

FileSystem Module

ProcessManagement

ProcessManagement Module

RunTime

RunTime Module

Track

Track Module

Vine

Vine gRPC Module

WorkSpace

WorkSpace Module