Mountain

Module Environment

Environment Module

Provides the concrete implementation of the application’s Environment.

This module contains the MountainEnvironment struct and all of its implementations of the provider traits defined in the Common crate. Each provider implementation is organized into its own file for clarity and separation of concerns.

Modules

CommandProvider

CommandProvider Implementation

ConfigurationProvider

ConfigurationProvider Implementation

CustomEditorProvider

CustomEditorProvider Implementation

DebugProvider

DebugProvider Implementation

DiagnosticProvider

DiagnosticProvider Implementation

DocumentProvider

DocumentProvider Implementation

FileSystemProvider

FileSystemProvider Implementation

IPCProvider

IPCProvider Implementation

KeybindingProvider

KeybindingProvider Implementation

LanguageFeatureProvider

LanguageFeatureProvider Implementation

MountainEnvironment

MountainEnvironment

OutputProvider

OutputProvider Implementation

SearchProvider

SearchProvider Implementation

SecretProvider

SecretProvider Implementation

SourceControlManagementProvider

SourceControlManagementProvider Implementation

StatusBarProvider

StatusBarProvider Implementation

StorageProvider

StorageProvider Implementation

SynchronizationProvider

SynchronizationProvider Implementation

TerminalProvider

TerminalProvider Implementation

TestProvider

TestProvider Implementation

TreeViewProvider

TreeViewProvider Implementation

UserInterfaceProvider

UserInterfaceProvider Implementation

Utility

Environment Utility

WebViewProvider

WebViewProvider Implementation

WorkSpaceProvider

WorkSpaceProvider Implementation