Mountain::Vine::Generated::cocoon_service_client

Struct CocoonServiceClient

pub struct CocoonServiceClient<T> {
    inner: Grpc<T>,
}

Service running on the Cocoon sidecar, listening for requests from Mountain.

Fields
inner: Grpc<T>
Implementations

impl CocoonServiceClient<Channel>

pub async fn connect<D>(dst: D) -> Result<Self, Error>
where D: TryInto<Endpoint>, D::Error: Into<StdError>,

Attempt to create a new client by connecting to a given endpoint.

impl<T> CocoonServiceClient<T>
where T: GrpcService<Body>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F, ) -> CocoonServiceClient<InterceptedService<T, F>>
where F: Interceptor, T::ResponseBody: Default, T: Service<Request<Body>, Response = Response<<T as GrpcService<Body>>::ResponseBody>>, <T as Service<Request<Body>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn process_mountain_request( &mut self, request: impl IntoRequest<GenericRequest>, ) -> Result<Response<GenericResponse>, Status>

A generic request-response method for Mountain to call a function on Cocoon.

pub async fn send_mountain_notification( &mut self, request: impl IntoRequest<GenericNotification>, ) -> Result<Response<Empty>, Status>

A generic fire-and-forget method for Mountain to send a notification to Cocoon.

pub async fn cancel_operation( &mut self, request: impl IntoRequest<CancelOperationRequest>, ) -> Result<Response<Empty>, Status>

A method for Mountain to request that Cocoon cancel a long-running operation.

pub async fn initial_handshake( &mut self, request: impl IntoRequest<Empty>, ) -> Result<Response<Empty>, Status>

Handshake - Called by Cocoon to signal readiness

pub async fn init_extension_host( &mut self, request: impl IntoRequest<InitExtensionHostRequest>, ) -> Result<Response<Empty>, Status>

Initialize Extension Host - Mountain sends initialization data to Cocoon

pub async fn register_command( &mut self, request: impl IntoRequest<RegisterCommandRequest>, ) -> Result<Response<Empty>, Status>

Register Command - Cocoon registers an extension command

pub async fn execute_contributed_command( &mut self, request: impl IntoRequest<ExecuteCommandRequest>, ) -> Result<Response<ExecuteCommandResponse>, Status>

Execute Contributed Command - Mountain executes an extension command

pub async fn unregister_command( &mut self, request: impl IntoRequest<UnregisterCommandRequest>, ) -> Result<Response<Empty>, Status>

Unregister Command - Unregister a previously registered command

pub async fn register_hover_provider( &mut self, request: impl IntoRequest<RegisterProviderRequest>, ) -> Result<Response<Empty>, Status>

Register Hover Provider - Register a hover provider

pub async fn provide_hover( &mut self, request: impl IntoRequest<ProvideHoverRequest>, ) -> Result<Response<ProvideHoverResponse>, Status>

Provide Hover - Request hover information

pub async fn register_completion_item_provider( &mut self, request: impl IntoRequest<RegisterProviderRequest>, ) -> Result<Response<Empty>, Status>

Register Completion Item Provider - Register a completion provider

pub async fn provide_completion_items( &mut self, request: impl IntoRequest<ProvideCompletionItemsRequest>, ) -> Result<Response<ProvideCompletionItemsResponse>, Status>

Provide Completion Items - Request completion items

pub async fn register_definition_provider( &mut self, request: impl IntoRequest<RegisterProviderRequest>, ) -> Result<Response<Empty>, Status>

Register Definition Provider - Register a definition provider

pub async fn provide_definition( &mut self, request: impl IntoRequest<ProvideDefinitionRequest>, ) -> Result<Response<ProvideDefinitionResponse>, Status>

Provide Definition - Request definition location

pub async fn register_reference_provider( &mut self, request: impl IntoRequest<RegisterProviderRequest>, ) -> Result<Response<Empty>, Status>

Register Reference Provider - Register a reference provider

pub async fn provide_references( &mut self, request: impl IntoRequest<ProvideReferencesRequest>, ) -> Result<Response<ProvideReferencesResponse>, Status>

Provide References - Request references

pub async fn register_code_actions_provider( &mut self, request: impl IntoRequest<RegisterProviderRequest>, ) -> Result<Response<Empty>, Status>

Register Code Actions Provider - Register code actions provider

pub async fn provide_code_actions( &mut self, request: impl IntoRequest<ProvideCodeActionsRequest>, ) -> Result<Response<ProvideCodeActionsResponse>, Status>

Provide Code Actions - Request code actions

pub async fn show_text_document( &mut self, request: impl IntoRequest<ShowTextDocumentRequest>, ) -> Result<Response<ShowTextDocumentResponse>, Status>

Show Text Document - Open a text document

pub async fn show_information_message( &mut self, request: impl IntoRequest<ShowMessageRequest>, ) -> Result<Response<ShowMessageResponse>, Status>

Show Information Message - Display an info message

pub async fn show_warning_message( &mut self, request: impl IntoRequest<ShowMessageRequest>, ) -> Result<Response<ShowMessageResponse>, Status>

Show Warning Message - Display a warning message

pub async fn show_error_message( &mut self, request: impl IntoRequest<ShowMessageRequest>, ) -> Result<Response<ShowMessageResponse>, Status>

Show Error Message - Display an error message

pub async fn create_status_bar_item( &mut self, request: impl IntoRequest<CreateStatusBarItemRequest>, ) -> Result<Response<CreateStatusBarItemResponse>, Status>

Create Status Bar Item - Create a status bar item

pub async fn set_status_bar_text( &mut self, request: impl IntoRequest<SetStatusBarTextRequest>, ) -> Result<Response<Empty>, Status>

Set Status Bar Text - Set status bar text

pub async fn create_webview_panel( &mut self, request: impl IntoRequest<CreateWebviewPanelRequest>, ) -> Result<Response<CreateWebviewPanelResponse>, Status>

Create Webview Panel - Create a new webview panel

pub async fn set_webview_html( &mut self, request: impl IntoRequest<SetWebviewHtmlRequest>, ) -> Result<Response<Empty>, Status>

Set Webview HTML - Update webview HTML content

pub async fn on_did_receive_message( &mut self, request: impl IntoRequest<OnDidReceiveMessageRequest>, ) -> Result<Response<Empty>, Status>

On Did Receive Message - Receive message from webview

pub async fn read_file( &mut self, request: impl IntoRequest<ReadFileRequest>, ) -> Result<Response<ReadFileResponse>, Status>

Read File - Read file contents

pub async fn write_file( &mut self, request: impl IntoRequest<WriteFileRequest>, ) -> Result<Response<Empty>, Status>

Write File - Write file contents

pub async fn stat( &mut self, request: impl IntoRequest<StatRequest>, ) -> Result<Response<StatResponse>, Status>

Stat - Get file metadata

pub async fn readdir( &mut self, request: impl IntoRequest<ReaddirRequest>, ) -> Result<Response<ReaddirResponse>, Status>

Read Directory - List directory contents

pub async fn watch_file( &mut self, request: impl IntoRequest<WatchFileRequest>, ) -> Result<Response<Empty>, Status>

Watch File - Watch file for changes

pub async fn find_files( &mut self, request: impl IntoRequest<FindFilesRequest>, ) -> Result<Response<FindFilesResponse>, Status>

Find Files - Search for files

pub async fn find_text_in_files( &mut self, request: impl IntoRequest<FindTextInFilesRequest>, ) -> Result<Response<FindTextInFilesResponse>, Status>

Find Text in Files - Search for text across files

pub async fn open_document( &mut self, request: impl IntoRequest<OpenDocumentRequest>, ) -> Result<Response<OpenDocumentResponse>, Status>

Open Document - Open a document

pub async fn save_all( &mut self, request: impl IntoRequest<SaveAllRequest>, ) -> Result<Response<SaveAllResponse>, Status>

Save All - Save all open documents

pub async fn apply_edit( &mut self, request: impl IntoRequest<ApplyEditRequest>, ) -> Result<Response<ApplyEditResponse>, Status>

Apply Edit - Apply a text edit to a document

pub async fn update_configuration( &mut self, request: impl IntoRequest<UpdateConfigurationRequest>, ) -> Result<Response<Empty>, Status>

Update Configuration - Notify of configuration changes

pub async fn update_workspace_folders( &mut self, request: impl IntoRequest<UpdateWorkspaceFoldersRequest>, ) -> Result<Response<Empty>, Status>

Update Workspace Folders - Update workspace folders

pub async fn open_terminal( &mut self, request: impl IntoRequest<OpenTerminalRequest>, ) -> Result<Response<Empty>, Status>

Open Terminal - Open a new terminal

pub async fn terminal_input( &mut self, request: impl IntoRequest<TerminalInputRequest>, ) -> Result<Response<Empty>, Status>

Terminal Input - Send input to terminal

pub async fn close_terminal( &mut self, request: impl IntoRequest<CloseTerminalRequest>, ) -> Result<Response<Empty>, Status>

Close Terminal - Close a terminal

pub async fn accept_terminal_opened( &mut self, request: impl IntoRequest<TerminalOpenedNotification>, ) -> Result<Response<Empty>, Status>

Accept Terminal Opened - Notification: Terminal opened

pub async fn accept_terminal_closed( &mut self, request: impl IntoRequest<TerminalClosedNotification>, ) -> Result<Response<Empty>, Status>

Accept Terminal Closed - Notification: Terminal closed

pub async fn accept_terminal_process_id( &mut self, request: impl IntoRequest<TerminalProcessIdNotification>, ) -> Result<Response<Empty>, Status>

Accept Terminal Process ID - Notification: Terminal process ID

pub async fn accept_terminal_process_data( &mut self, request: impl IntoRequest<TerminalDataNotification>, ) -> Result<Response<Empty>, Status>

Accept Terminal Process Data - Notification: Terminal output

pub async fn register_tree_view_provider( &mut self, request: impl IntoRequest<RegisterTreeViewProviderRequest>, ) -> Result<Response<Empty>, Status>

Register Tree View Provider - Register a tree view provider

pub async fn get_tree_children( &mut self, request: impl IntoRequest<GetTreeChildrenRequest>, ) -> Result<Response<GetTreeChildrenResponse>, Status>

Get Tree Children - Request tree view children

pub async fn register_scm_provider( &mut self, request: impl IntoRequest<RegisterScmProviderRequest>, ) -> Result<Response<Empty>, Status>

Register SCM Provider - Register source control provider

pub async fn update_scm_group( &mut self, request: impl IntoRequest<UpdateScmGroupRequest>, ) -> Result<Response<Empty>, Status>

Update SCM Group - Update SCM group

pub async fn git_exec( &mut self, request: impl IntoRequest<GitExecRequest>, ) -> Result<Response<GitExecResponse>, Status>

Execute Git - Execute git command

pub async fn register_debug_adapter( &mut self, request: impl IntoRequest<RegisterDebugAdapterRequest>, ) -> Result<Response<Empty>, Status>

Register Debug Adapter - Register debug adapter

pub async fn start_debugging( &mut self, request: impl IntoRequest<StartDebuggingRequest>, ) -> Result<Response<StartDebuggingResponse>, Status>

Start Debugging - Start debug session

pub async fn participate_in_save( &mut self, request: impl IntoRequest<ParticipateInSaveRequest>, ) -> Result<Response<ParticipateInSaveResponse>, Status>

Participate in Save - Extension participates in save

pub async fn get_secret( &mut self, request: impl IntoRequest<GetSecretRequest>, ) -> Result<Response<GetSecretResponse>, Status>

Get Secret - Retrieve a secret from storage

pub async fn store_secret( &mut self, request: impl IntoRequest<StoreSecretRequest>, ) -> Result<Response<Empty>, Status>

Store Secret - Store a secret in storage

pub async fn delete_secret( &mut self, request: impl IntoRequest<DeleteSecretRequest>, ) -> Result<Response<Empty>, Status>

Delete Secret - Delete a secret from storage